[ALPS04016653] DISP: fix build error
[Detail]
fix build error in ARM 32bit project

MTK-Commit-Id: dae96ecf72ce9ff14e79950f9fbde12ec0ba2159

CR-Id: ALPS04016653
Feature: Display Driver (DSI Interface)
(cherry-pick from 07017bf49b196e23fa67265b673f1bb8979df033)

+++ drivers/misc/mediatek/video/mt6771/videox/debug.c
@@ -76,13 +76,13 @@ int fps_show_flag;
 struct fps_debug fps_info_debug = {0};
 
 /* hrt */
-int hrt_high, hrt_low;
+unsigned long long hrt_high, hrt_low;
 int hrt_show_flag;
 
 /* layer_en */
 int layer_en_num;
 int layer_en_num_flag;
-int layer_size_high;
+uint64_t layer_size_high;
 int layer_size_low;
 int layer_size_flag;
